requirements.txt
flask==2.0.1
werkzeug==2.0.3
openai==0.27.0
python-dotenv==0.19.0
reportlab==4.0.4
markdown==3.3.4 